scripts/requirements.txt
pandas==2.1.4
numpy==1.24.3
torch==2.1.1
transformers==4.36.0
faiss-cpu==1.7.4
tqdm==4.66.1
scikit-learn==1.3.2

